Refactor RFC4.java and Method.java
